Split off connection
split off the connectToDatabase() function to give access to other pages.
